Created new SpendingPage to let user enter amount spent, added amount spent to total amount in firebase, and loaded all values in budgetpage pie chart.

+++ Coding Project/Money-Man/app/src/main/java/com/example/money_man_group1/userSpendingInfo.kt
@@ -15,7 +15,7 @@ data class UserSpendingInfo(
         var isCategoryActive: String = "No",
         var categoryName: String = "",
         var categoryDescription: String = "",
-        var maxSpendingLimit: Int = 0,
-        var currentMoneySpent: Int = 0
+        var maxSpendingLimit: Double = 0.0,
+        var currentMoneySpent: Double = 0.0
     )
 }
